fix get completed tasks endpoint

+++ src/main/java/com/lans/fleems/statistic/StatisticController.java
@@ -38,7 +38,7 @@ public class StatisticController {
     }
 
     @GetMapping("/completed")
-    public ResponseEntity<Double[]> getCompletedThisYear() {
+    public ResponseEntity<int[]> getCompletedThisYear() {
         return ResponseEntity.ok(statisticService.getCompletedThisYear());
     }
 }
